+++ src/builtin/builtin_exit.c
@@ -6,7 +6,7 @@
 /*   By: hujeong <hujeong@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 16:18:52 by migo              #+#    #+#             */
-/*   Updated: 2023/03/10 10:46:13 by hujeong          ###   ########.fr       */
+/*   Updated: 2023/03/10 14:05:40 by hujeong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,6 @@ void	not_num(char *str)
 
 void	builtin_exit(t_cmd *cmd, int count)
 {
-
 	if (count == 1)
 		write(1, "exit\n", 5);
 	if (cmd->option[1] == NULL)
