+++ src/builtin/builtin_pwd.c
@@ -6,7 +6,7 @@
 /*   By: hujeong <hujeong@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 20:19:53 by hujeong           #+#    #+#             */
-/*   Updated: 2023/03/10 10:58:55 by hujeong          ###   ########.fr       */
+/*   Updated: 2023/03/10 14:05:46 by hujeong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,4 +32,3 @@ int	builtin_pwd(t_current *current)
 	write(1, "\n", 1);
 	return (0);
 }
-
